% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQ_const.R
\name{EOQ_const}
\alias{EOQ_const}
\title{EOQ Constant Demand}
\usage{
EOQ_const(d, k, h, c)
}
\arguments{
\item{d}{double - constant demand in a cycle T (units/time).}

\item{k}{double - fixed cost for each order ($/order).}

\item{h}{double - holding cost for each unit per time ($/(unit*time)).}

\item{c}{double - purchase cost for each unit ($/unit).

All time units must be the same (year, month, day, etc.).}
}
\value{
double - Economic Order Quantity Q*
}
\description{
This function gets the optimal EOQ orders with constant demand in one cycle T.
}
\examples{
EOQ_const(400,100,0.4,200)
}
