% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQ_ncd.R
\name{EOQ_ncd}
\alias{EOQ_ncd}
\title{EOQ no Constant Demand}
\usage{
EOQ_ncd(n, d, t, k, h, c)
}
\arguments{
\item{n}{int - number of different demand rates in a cycle T.}

\item{d}{vector - vector of different demand rates in a cycle T (u/time). The vector length must be equal to @param n.}

\item{t}{vector - vector of percentages of cycle for each demand rate in vector d. The sum of this vector should be 1. The vector length must be equal to @param n.}

\item{k}{double - fixed cost for each order ($/order).}

\item{h}{double - holding cost for each unit per time ($/(u*time)).}

\item{c}{double - purchase cost for each unit ($/u).}
}
\value{
double
}
\description{
This function gets the optimal EOQ orders with different deterministic demand rates in one cycle T.
}
\details{
All time units must be the same (year, month, day, etc.).
}
\examples{
EOQ_ncd(3, c(6.75, 0, 9), c(0.5, 0.3, 0.2), 1700, 4.5, 35)
}
