% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQ_tknown.R
\name{EOQ_tknown}
\alias{EOQ_tknown}
\title{EOQ with Different Demand Rates in a Time of Cycle Known}
\usage{
EOQ_tknown(n, d, t, k, h, c)
}
\arguments{
\item{n}{int - number of different demand rates in a cycle T.}

\item{d}{vector - vector of different demand rates in a cycle T (u/time). The vector length must be equal to @param n.}

\item{t}{vector - vector of different times (time) in a cycle T for each demand rate in vector d. The sum of this vector should be equal to T. The vector length must be equal to @param n.}

\item{k}{double - fixed cost for each order ($/order).}

\item{h}{double - holding cost for each unit per time ($/(u*time)).}

\item{c}{double - purchase cost for each unit ($/u).}
}
\value{
double - Economic Order Quantity Q*
}
\description{
This function gets the optimal EOQ orders with different deterministic demand rates in one cycle T, where this time is known with different sub-times (seasonal).
}
\details{
All time units must be the same (year, month, day, etc.).
}
\examples{
EOQ_tknown(5,c(150,450,0,300,0),c(12,5,2,4,1),1500,5,17)
}
