% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrNormal.R
\name{qrNormal}
\alias{qrNormal}
\title{QR with Normal Distribution Demand}
\usage{
qrNormal(m, s, k, c, h, p, t, b)
}
\arguments{
\item{m}{double - mean of the demand (u/time).}

\item{s}{double - standard deviation of the demand (u/time).}

\item{k}{double - fixed cost for each order ($/order).}

\item{c}{double - purchase cost for each unit ($/u).}

\item{h}{double - holding cost for each unit per time ($/(u*time)).}

\item{p}{double - cost for each unit shortage during Lead Time ($/(u missing))}

\item{t}{double - Lead Time, time an order takes to arrive once is placed (time)}

\item{b}{boolean - TRUE if the system manages Backorders, FALSE if the system manages Lost Sales}
}
\value{
vector - c(Q,R)
}
\description{
This function gets the optimal order quantity Q and the reorder point R with a Normal Distribution Demand in a Continuous Review System.
}
\details{
All time units must be the same (year, month, day, etc.).
}
\examples{
qrNormal(10,2.83,32,5,4,10,0.5,TRUE)
}
